# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Herbaflow API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Create Analysis",
        "operationId": "create_analysis",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "List Analyses",
        "operationId": "list_analyses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisListItem"
                  },
                  "title": "Response List Analyses"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "analyses"
        ],
        "summary": "Delete Analysis",
        "operationId": "delete_analysis",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/advance": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Advance Analysis",
        "operationId": "advance_analysis",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/reset-from/{stage}": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Reset From",
        "operationId": "reset_from",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stage"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetFromRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/stages/{stage}/edit": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Edit Stage",
        "operationId": "edit_stage",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stage"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageEditRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisCreate": {
        "properties": {
          "analysis_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Name"
          },
          "plant_input_mode": {
            "$ref": "#/components/schemas/PlantInputMode",
            "default": "selection"
          },
          "disease_input_mode": {
            "$ref": "#/components/schemas/DiseaseInputMode",
            "default": "selection"
          },
          "plant_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Plant Ids"
          },
          "disease_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Id"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode",
            "default": "guided"
          },
          "manual_compound_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Manual Compound Ids"
          },
          "manual_target_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Manual Target Ids"
          },
          "manual_disease_target_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Manual Disease Target Ids"
          },
          "plant_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Label"
          },
          "disease_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Label"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters",
            "description": "Per-group parameter overrides for any pipeline group (adme / target / disease_targets / ppi / hub_genes / enrichment); validated against the contract hard bounds at create, merged over the group defaults, and frozen as the run baseline."
          }
        },
        "type": "object",
        "title": "AnalysisCreate"
      },
      "AnalysisListItem": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "analysis_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "disease_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Id"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "analysis_name",
          "status",
          "current_stage",
          "created_at",
          "completed_at",
          "disease_id"
        ],
        "title": "AnalysisListItem",
        "description": "Lean per-run shape for the recent-runs list.\n\nDeliberately omits ``stage_results`` (the heavy per-stage tables).\n``parameters`` is kept \u2014 it carries ``input_modes`` / ``labels`` / ``plant_ids``;\nits size is bounded by the entity caps."
      },
      "AnalysisRead": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "analysis_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Name"
          },
          "disease_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Id"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "stage_results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stage Results"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "stage_state": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Stage State",
            "description": "Per-stage entry-mode state (``computed``/``user_provided``/``not_applicable``) derived\nfrom the run's ``input_modes`` via the canonical matrix. This is the entry-mode-INTENDED\nstate for the FE (greyed N/A stages + the \"provided by you\" badge), distinct from the\npresentational ``stage_results[n].state`` \u2014 which the durable edit layer also flips to\n``user_provided`` when a *computed* entity stage is merely edited. A run with no\n``input_modes`` (legacy / all-selection) yields all ``computed``.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "analysis_name",
          "disease_id",
          "mode",
          "status",
          "current_stage",
          "stage_results",
          "created_at",
          "completed_at",
          "expires_at",
          "error_message",
          "stage_state"
        ],
        "title": "AnalysisRead"
      },
...
      "DiseaseInputMode": {
        "type": "string",
        "enum": [
          "selection",
          "manual_disease_targets"
        ],
        "title": "DiseaseInputMode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Mode": {
        "type": "string",
        "enum": [
          "auto",
          "guided"
        ],
        "title": "Mode"
      },
      "PlantInputMode": {
        "type": "string",
        "enum": [
          "selection",
          "manual_compounds",
          "manual_targets"
        ],
        "title": "PlantInputMode"
      },
...
      "ResetFromRequest": {
        "properties": {
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          }
        },
        "type": "object",
        "title": "ResetFromRequest",
        "description": "Body for POST /analyses/{id}/reset-from/{stage}.\n\n``parameters`` is a stage-keyed map of ADME param overrides, e.g.\n``{\"2\": {\"max_violations\": 0}}``.  Only the entry matching the target\nstage is extracted and forwarded to the service."
      },
...
      "StageEditRequest": {
        "properties": {
          "add": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Add"
          },
          "remove": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Remove"
          }
        },
        "type": "object",
        "title": "StageEditRequest",
        "description": "Body for POST /analyses/{id}/stages/{stage}/edit."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}